Changed average rating into a float

+++ backend/server.py
@@ -82,7 +82,7 @@ def show_game_information(game_id):
     ave_score = crud.get_average_rating_by_id(game_id)
 
     if ave_score != 'No ratings for this game':
-        game['score'] = int(ave_score)
+        game['score'] = float(ave_score)
 
     return jsonify(game)
